bcm63xx-ast-chan : add new parameters 'dialing_timeout_1st_digit' and 'search_extension_trigger' + fix a bug in bcmph_chan_write().
Following a suggestion of user stesc, pointing out that generally the delay between the user hooks off the phone and dials the first digit is longer than the delay between the user dials the other digits, there's a new parameter called 'dialing_timeout_1st_digit'.
It's the time the channel waits before searching the default extension 's', after the users hook off and if no digit is dialed.
As soon as the user dialed a digit, the channel uses the other timeout 'dialing_timeout' before searching if digits dialed form a valid extension.

The other new parameter 'search_extension_trigger' is a character that when dialed, triggers the search for a valid extension immediately, without waiting for 'dialing_timeout'.

Fix a bug in bcmph_chan_write that prevent to write data when call not answered.